creating a shared code for test
