+++ src/db/db-mutation.html
@@ -18,7 +18,7 @@
             </div>
             <br>
             <div class="controls">
-                <button id="upload">Upload</button>
+                <button id="upload">Import</button>
                 <button id="clear">Clear</button>
             </div>
         </div>
@@ -31,7 +31,7 @@
         <div class="vl"></div>
         <div class="delete-container">
             <h3>Click below to clear database (db is not deleted, but emptied).</h3>
-            <button id="delete">Delete</button>
+            <button id="delete">Reset</button>
         </div>
     </div>
     <hr>
